Use the 'short' Speech-to-Text model for push-to-talk transcription
'long' is tuned for continuous/longer audio and can bias toward the
wrong word continuation on a short single-sentence clip (e.g. 今日 heard
as 企業). 'short' matches this route's actual use case.

+++ src/lib/googleStt.ts
@@ -49,7 +49,10 @@ export async function transcribeSpeech(audioBuffer: Buffer, languageCode: string
           config: {
             autoDecodingConfig: {},
             languageCodes: [languageCode],
-            model: 'long',
+            // 'short' is tuned for brief, single-utterance audio (a few seconds) — matches this
+            // route's push-to-talk use case better than 'long', which targets continuous/longer
+            // recordings and can bias toward different word continuations on a short clip.
+            model: 'short',
           },
           content: audioBuffer.toString('base64'),
         }),
